pw_hdlc: Multibyte address support
This extends the HDLC encoder and decoder in both C++ and Python to use
HDLC extended addressing, where the address field is a one-terminated
LSB varint.

Classes using HDLC, such as the HDLC ChannelOutput and WirePacketParser,
are updated accordingly.

Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/38020

+++ pw_hdlc/public/pw_hdlc/encoder.h
@@ -20,16 +20,16 @@
 namespace pw::hdlc {
 
 // Writes an HDLC unnumbered information frame (UI-frame) to the provided
-// writer. The frame contains the following:
+// writer. The complete frame contains the following:
 //
 //   - HDLC flag byte (0x7e)
 //   - Address
-//   - UI-frame control byte
-//   - Data (0 or more bytes)
+//   - UI-frame control (metadata) byte
+//   - Payload (0 or more bytes)
 //   - Frame check sequence (CRC-32)
 //   - HDLC flag byte (0x7e)
 //
-Status WriteUIFrame(uint8_t address,
+Status WriteUIFrame(uint64_t address,
                     ConstByteSpan payload,
                     stream::Writer& writer);
